Restructure pass one complete!
