+++ packages/core/src/Renderer.js
@@ -1,17 +1,17 @@
 import AbstractRenderer from './AbstractRenderer';
 import { sayHello } from '@pixi/utils';
-import MaskSystem from './systems/MaskSystem';
-import StencilSystem from './systems/StencilSystem';
+import MaskSystem from './mask/MaskSystem';
+import StencilSystem from './mask/StencilSystem';
 import FilterSystem from './filters/FilterSystem';
-import FramebufferSystem from './systems/FramebufferSystem';
-import RenderTextureSystem from './systems/RenderTextureSystem';
+import FramebufferSystem from './framebuffer/FramebufferSystem';
+import RenderTextureSystem from './renderTexture/RenderTextureSystem';
 import TextureSystem from './textures/TextureSystem';
-import ProjectionSystem from './systems/ProjectionSystem';
-import StateSystem from './systems/StateSystem';
+import ProjectionSystem from './projection/ProjectionSystem';
+import StateSystem from './state/StateSystem';
 import GeometrySystem from './geometry/GeometrySystem';
 import ShaderSystem from './shader/ShaderSystem';
-import ContextSystem from './systems/ContextSystem';
-import BatchSystem from './systems/BatchSystem';
+import ContextSystem from './context/ContextSystem';
+import BatchSystem from './batch/BatchSystem';
 import TextureGCSystem from './textures/TextureGCSystem';
 import { RENDERER_TYPE } from '@pixi/constants';
 import UniformGroup from './shader/UniformGroup';
