+++ packages/core/src/batch/ObjectRenderer.js
@@ -1,4 +1,4 @@
-import WebGLSystem from '../systems/WebGLSystem';
+import WebGLSystem from '../WebGLSystem';
 
 /**
  * Base for a common object renderer that can be used as a system renderer plugin.
