+++ packages/core/src/geometry/GeometrySystem.js
@@ -1,4 +1,4 @@
-import WebGLSystem from '../systems/WebGLSystem';
+import WebGLSystem from '../WebGLSystem';
 import GLBuffer from './GLBuffer';
 
 const byteSizeMap = { 5126: 4, 5123: 2, 5121: 1 };
