+++ packages/core/src/mask/StencilSystem.js
@@ -1,4 +1,4 @@
-import WebGLSystem from './WebGLSystem';
+import WebGLSystem from '../WebGLSystem';
 
 /**
  * @class
