+++ packages/core/src/shader/utils/checkMaxIfStatmentsInShader.js
