--- packages/core/src/systems/index.js
@@ -1,5 +1,0 @@
-/**
- * Systems are individual components to the Renderer pipeline.
- * @namespace PIXI.systems
- */
-export { default as WebGLSystem } from './WebGLSystem';
